FossilOrigin-Name: be32067a2ac7b7d7fd586ddbea91450d22ed9996e66116c43d0d3c999a3ad932

+++ src/vdbeInt.h
@@ -393,7 +393,7 @@ struct sqlite3_context {
   int isError;            /* Error code returned by the function. */
   u8 enc;                 /* Encoding to use for results */
   u8 skipFlag;            /* Skip accumulator loading if true */
-  u8 argc;                /* Number of arguments */
+  u16 argc;               /* Number of arguments */
   sqlite3_value *argv[1]; /* Argument set */
 };
